requirements.txt
speedtest-cli==2.1.3
tqdm==4.66.5






